refactor: RacingController의 기능을 ViewController를 생성해 분할
