+++ src/main/kotlin/racingcar/utils/Validator.kt
@@ -29,4 +29,12 @@ class Validator {
             throw IllegalArgumentException(CAR_NAME_BOUNDARY_ERROR_MESSAGE)
         }
     }
+
+    companion object {
+        const val CAR_NAME_BOUNDARY_ERROR_MESSAGE = "자동차 이름의 길이는 $MIN_CAR_NAME_LENGTH 이상 $MAX_CAR_NAME_LENGTH 이하로 부탁이요~"
+        const val DUPLICATED_CAR_NAME_ERROR_MESSAGE = "자동차 이름 중복 빼주세요~"
+
+        const val NOT_INTEGER_TYPE_ERROR_MESSAGE = "숫자 형태로 부탁이요~"
+        const val ROUND_COUNT_BOUNDARY_ERROR_MESSAGE = "라운드 횟수는 $MIN_ROUND_COUNT 이상 $MAX_ROUND_COUNT 이하로 부탁이요~"
+    }
 }
