+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,13 +1,13 @@
 package racingcar.view
 
-import racingcar.utils.CAR_NAME_DELIMITER
+import racingcar.utils.TEXT_IN_LINE_DELIMITER
 import racingcar.utils.Validator
 import racingcar.utils.TextUtils
 
 class InputView(private val validator: Validator) {
 
     fun readCarNames(): List<String> {
-        var names = readln().split(CAR_NAME_DELIMITER).toList()
+        var names = readln().split(TEXT_IN_LINE_DELIMITER).toList()
         names = TextUtils.removeTextsBlank(names)
 
         return validator.checkCarNames(names)
