+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -1,7 +1,6 @@
 package racingcar.view
 
 import racingcar.model.Car
-import racingcar.utils.WINNER_NOTIFICATION_MESSAGE
 
 class OutputView {
     fun printMessage(message: String) = println(message)
@@ -11,9 +10,5 @@ class OutputView {
             println("$car : ${car.getPositionAsDash()}")
         }
         println()
-    }
-
-    fun printWinners(winners: List<Car>) {
-        println("$WINNER_NOTIFICATION_MESSAGE: ${winners.joinToString(", ")}")
     }
 }
